requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-mysql==1.5.2
paho-mqtt==1.6.1
flask-jwt-extended==4.5.2
flask_swagger_ui==4.11.1